TIP-860: Move FamilyVariant related api tests
